+++ structs/NBAPlayerProgressions.go
@@ -1,17 +1,17 @@
 package structs
 
 type NBAPlayerProgressions struct {
-	Age              int
-	Overall          int
-	Shooting2        int
-	Shooting3        int
-	FreeThrow        int
-	Finishing        int
-	Ballwork         int
-	Rebounding       int
-	InteriorDefense  int
-	PerimeterDefense int
-	Stamina          int
+	Age              uint8
+	Overall          uint8
+	Shooting2        uint8
+	Shooting3        uint8
+	FreeThrow        uint8
+	Finishing        uint8
+	Ballwork         uint8
+	Rebounding       uint8
+	InteriorDefense  uint8
+	PerimeterDefense uint8
+	Stamina          uint8
 	IsRetiring       bool
 	PotentialGrade   string
 }
